+++ mm/2s2h/Enhancements/Controls/Mouse/Mouse.cpp
@@ -6,8 +6,6 @@
 #include "public/bridge/consolevariablebridge.h"
 #include "overlays/kaleido_scope/ovl_kaleido_scope/z_kaleido_scope.h"
 
-#define MOUSE_ENABLED (Mouse_IsCaptured() && CVarGetInteger("gEnhancements.Camera.Mouse.Enabled", 0))
-
 static MouseCoords current;
 
 #ifdef __cplusplus
@@ -37,48 +35,6 @@ bool Mouse_IsCaptured() {
     return Ship::Context::GetInstance()->GetWindow()->IsMouseCaptured();
 }
 
-static s8 iterMouse = 0;
-static f32 mouseQuickspinX[5] = {};
-static f32 mouseQuickspinY[5] = {};
-static u8 quickspinCount = 0;
-
-// copypasted logic from SoH
-void UpdateQuickspinCount(Input* input) {
-    quickspinCount = (quickspinCount + 1) % 5;
-    mouseQuickspinX[quickspinCount] = current.x;
-    mouseQuickspinY[quickspinCount] = current.y;
-}
-
-bool HandleQuickspin(bool* should, s8* iter2, s8* sp3C) {
-    s8 temp1;
-    s8 temp2;
-    s32 i;
-
-    for (i = 0; i < 4; i++, iter2++) {
-        f32 relY = mouseQuickspinY[i + 1] - mouseQuickspinY[i];
-        f32 relX = mouseQuickspinX[i + 1] - mouseQuickspinX[i];
-        s16 aTan = Math_Atan2S(relY, -relX);
-        iterMouse = (u16)(aTan + 0x2000) >> 9;
-        if ((*iter2 = iterMouse) < 0) {
-            return *should = false;
-        }
-        *iter2 *= 2;
-    }
-    temp1 = sp3C[0] - sp3C[1];
-    if (ABS(temp1) < 10) {
-        return *should = false;
-    }
-    iter2 = &sp3C[1];
-    for (i = 1; i < 3; i++, iter2++) {
-        temp2 = *iter2 - *(iter2 + 1);
-        if ((ABS(temp2) < 10) || (temp2 * temp1 < 0)) {
-            return *should = false;
-        }
-    }
-
-    return *should = true;
-}
-
 void HandlePauseCapture(PauseContext* pauseCtx) {
     static bool buf = false;
     static bool paused = false;
@@ -104,17 +60,6 @@ void HandlePauseCapture(PauseContext* pauseCtx) {
 }
 
 void RegisterMouseRelatedHooks() {
-    COND_VB_SHOULD(
-        VB_SHOULD_QUICKSPIN,
-        MOUSE_ENABLED && CVarGetInteger("gEnhancements.Mouse.Quickspin.Enable", 0),
-        { HandleQuickspin(should, va_arg(args, s8*), va_arg(args, s8*)); }
-    );
-    COND_HOOK(
-        OnPassPlayerInputs,
-        MOUSE_ENABLED && CVarGetInteger("gEnhancements.Mouse.Quickspin.Enable", 0),
-        UpdateQuickspinCount
-    );
-
     COND_HOOK(
         OnKaleidoUpdate,
         true,
